+++ src/main/java/org/nebula/contrib/ngbatis/handler/NgSubgraphResultHandler.java
@@ -7,6 +7,7 @@ package org.nebula.contrib.ngbatis.handler;
 import com.vesoft.nebula.client.graph.data.ResultSet;
 import com.vesoft.nebula.client.graph.data.ResultSet.Record;
 import com.vesoft.nebula.client.graph.data.ValueWrapper;
+import java.util.ArrayList;
 import java.util.List;
 import org.nebula.contrib.ngbatis.models.data.NgEdge;
 import org.nebula.contrib.ngbatis.models.data.NgSubgraph;
@@ -36,7 +37,7 @@ public class NgSubgraphResultHandler extends AbstractResultHandler<NgSubgraph<?>
   public NgSubgraph<?> handle(NgSubgraph<?> newResult, Record row) {
     ValueWrapper nodes = row.get("nodes");
     if (nodes != null) {
-      List<ValueWrapper> nodesValueWrapper = nodes.asList();
+      List<ValueWrapper> nodesValueWrapper = getCollection(nodes);
       nodesValueWrapper.forEach(nodeValueWrapper -> {
         List vertexes = newResult.getVertexes();
         NgVertex<?> vertex = new NgVertex<>();
@@ -47,7 +48,7 @@ public class NgSubgraphResultHandler extends AbstractResultHandler<NgSubgraph<?>
 
     ValueWrapper relationships = row.get("relationships");
     if (relationships != null) {
-      List<ValueWrapper> edgesValueWrapper = relationships.asList();
+      List<ValueWrapper> edgesValueWrapper = getCollection(relationships);
       edgesValueWrapper.forEach(edgeValueWrapper -> {
         List edges = newResult.getEdges();
         NgEdge<?> edge = new NgEdge<>();
@@ -56,5 +57,16 @@ public class NgSubgraphResultHandler extends AbstractResultHandler<NgSubgraph<?>
       });
     }
     return newResult;
+  }
+
+  /**
+    * Supports both List and Set as param, 
+   *    and returns a List in order to be compatible with the original code.
+   * @param listOrSetWrapper original data from ResultSet
+   * @return List of ValueWrapper
+    */
+  private List<ValueWrapper> getCollection(ValueWrapper listOrSetWrapper) {
+    return listOrSetWrapper.isList() ? listOrSetWrapper.asList() 
+      : new ArrayList<>(listOrSetWrapper.asSet());
   }
 }